LPS-79327 Prior to this commit, I just wanted to make everything consistent. Now, my guess is that the word SegmentationSegment, while consistent, sounds horrible. So let's apply the Faro and Pulpo pattern. I found a word in Greek that means "part" or "segment". The word is "meris". See http://biblehub.com/greek/3310.htm It's short, and sounds beautiful to me. If you don't like the name, send me another. Should be super easy to do.
