+++ modules/apps/meris/meris-api/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Meris API
+Bundle-SymbolicName: com.liferay.meris.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.meris
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title:
