+++ modules/apps/meris/meris-api/src/main/resources/com/liferay/meris/packageinfo
