+++ modules/apps/meris/meris-asset-category-demo-test/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay Meris Asset Category Demo Test
+Bundle-SymbolicName: com.liferay.meris.asset.category.demo.test
+Bundle-Version: 1.0.0
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title:
