+++ modules/apps/meris/meris-asset-category-demo/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay Meris Asset Category Demo
+Bundle-SymbolicName: com.liferay.meris.asset.category.demo
+Bundle-Version: 1.0.0
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title:
