+++ modules/apps/meris/meris-asset-category-demo/src/main/java/com/liferay/meris/asset/category/demo/internal/AssetCategoryMerisSegment.java
@@ -1,0 +1,53 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.meris.asset.category.demo.internal;
+
+import com.liferay.asset.kernel.model.AssetCategory;
+import com.liferay.meris.MerisSegment;
+import com.liferay.portal.kernel.util.GetterUtil;
+
+import java.util.Locale;
+
+/**
+ * @author Eduardo Garcia
+ */
+public class AssetCategoryMerisSegment implements MerisSegment {
+
+	public AssetCategoryMerisSegment(AssetCategory assetCategory) {
+		_assetCategory = assetCategory;
+	}
+
+	public long getAssetCategoryId() {
+		return _assetCategory.getCategoryId();
+	}
+
+	@Override
+	public String getDescription(Locale locale) {
+		return _assetCategory.getDescription(locale);
+	}
+
+	@Override
+	public String getMerisSegmentId() {
+		return GetterUtil.getString(_assetCategory.getCategoryId());
+	}
+
+	@Override
+	public String getName(Locale locale) {
+		return _assetCategory.getTitle(locale);
+	}
+
+	private final AssetCategory _assetCategory;
+
+}
