--- modules/apps/segmentation/app.bnd
@@ -1,14 +1,0 @@
-Liferay-Releng-App-Description:
-Liferay-Releng-App-Title: ${liferay.releng.app.title.prefix} Segmentation
-Liferay-Releng-Bundle: true
-Liferay-Releng-Category: Productivity
-Liferay-Releng-Demo-Url:
-Liferay-Releng-Deprecated: false
-Liferay-Releng-Fix-Delivery-Method: core
-Liferay-Releng-Labs: false
-Liferay-Releng-Marketplace: true
-Liferay-Releng-Portal-Required: true
-Liferay-Releng-Public: ${liferay.releng.public}
-Liferay-Releng-Restart-Required: true
-Liferay-Releng-Support-Url: http://www.liferay.com
-Liferay-Releng-Supported: ${liferay.releng.supported}
