--- modules/apps/segmentation/segmentation-api/bnd.bnd
@@ -1,6 +1,0 @@
-Bundle-Name: Liferay Segmentation API
-Bundle-SymbolicName: com.liferay.segmentation.api
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.segmentation
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title:
