--- modules/apps/segmentation/segmentation-asset-category-demo-test/build.gradle
@@ -1,9 +1,0 @@
-sourceCompatibility = "1.8"
-targetCompatibility = "1.8"
-
-dependencies {
-	testIntegrationCompile group: "com.liferay", name: "com.liferay.arquillian.extension.junit.bridge", version: "1.0.7"
-	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
-	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	testIntegrationCompile project(":apps:segmentation:segmentation-api")
-}
