--- modules/apps/segmentation/segmentation-asset-category-demo/bnd.bnd
@@ -1,5 +1,0 @@
-Bundle-Name: Liferay Segmentation Asset Category Demo
-Bundle-SymbolicName: com.liferay.segmentation.asset.category.demo
-Bundle-Version: 1.0.0
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title:
