--- modules/apps/segmentation/segmentation-asset-category-demo/build.gradle
@@ -1,10 +1,0 @@
-sourceCompatibility = "1.8"
-targetCompatibility = "1.8"
-
-dependencies {
-	compileOnly group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
-	compileOnly project(":apps:segmentation:segmentation-api")
-	compileOnly project(":apps:static:osgi:osgi-util")
-}
